new file structure for more organisation
Splitting up from using just the singular app.py
