+++ backend/app/__init__.py
@@ -1,0 +1,11 @@
+from flask import Flask
+from flask_cors import CORS
+
+from .routes import setup_routes
+
+
+def create_app():
+    app = Flask(__name__)
+    CORS(app)
+    setup_routes(app)
+    return app
